optimize python watch dog
