+++ linkis-engineconn-plugins/spark/src/main/scala/org/apache/linkis/engineplugin/spark/executor/SparkPythonExecutor.scala
@@ -44,7 +44,6 @@ import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.apache.spark.sql.execution.datasources.csv.UDF
 
 import java.io._
-import java.lang.management.ManagementFactory
 import java.net.InetAddress
 import java.util
 
@@ -123,13 +122,13 @@ class SparkPythonExecutor(val sparkEngineSession: SparkEngineSession, val id: In
       Utils.tryAndErrorMsg {
         process.destroy()
         process = null
-
+        Thread.sleep(1000 * 2L)
         // process.destroy will kills the subprocess,not need to force kill with -9,
         // kill -9 may cause resources not to be released
-//        pid.foreach(p => {
-//          logger.info(s"skip Try to kill Pyspark process with: [kill -9 ${p}]")
-//           Utils.exec(Array("kill", "-9", p), 3000L)
-//        })
+        pid.foreach(p => {
+          logger.info(s"skip Try to kill Pyspark process with: [kill -9 ${p}]")
+          Utils.exec(Array("kill", "-9", p), 3000L)
+        })
 
       }("process close failed")
     }
@@ -211,7 +210,7 @@ class SparkPythonExecutor(val sparkEngineSession: SparkEngineSession, val id: In
     cmd.addArgument(pythonClasspath.toString(), false)
 
     // add java process pid
-    cmd.addArgument(getProcessId, false)
+    cmd.addArgument(Utils.getProcessId(), false)
 
     cmd.addArgument(pyFiles, false)
 
@@ -284,26 +283,6 @@ class SparkPythonExecutor(val sparkEngineSession: SparkEngineSession, val id: In
     lazyInitGateway()
     this.jobGroup = jobGroup
     executeLine(code)
-  }
-
-  // Note: may fail in some JVM implementations
-  private def getProcessId(): String = {
-    // therefore fallback has to be provided
-    // something like '<pid>@<hostname>', at least in SUN / Oracle JVMs
-    val jvmName = ManagementFactory.getRuntimeMXBean.getName
-    val index = jvmName.indexOf('@')
-    // part before '@' empty (index = 0) / '@' not found (index = -1)
-    if (index < 1) {
-      null
-    }
-    Utils.tryCatch {
-      val getpid = jvmName.substring(0, index)
-      logger.info(s"get java process Id:$getpid")
-      getpid
-    } { t =>
-      logger.info(s"Failed to get process Id with error", t.getMessage)
-      null
-    }
   }
 
   def lazyInitGateway(): Unit = {
